Conflicts:
	modules/ballerina-core/pom.xml
	modules/features/ballerina-core-feature/pom.xml
	pom.xml

+++ modules/ballerina-core/src/main/java/org/wso2/ballerina/core/nativeimpl/lang/json/GetDouble.java
@@ -43,7 +43,7 @@ import org.wso2.ballerina.core.nativeimpl.annotations.BallerinaFunction;
         functionName = "getDouble",
         args = {@Argument(name = "json", type = TypeEnum.JSON),
                 @Argument(name = "jsonPath", type = TypeEnum.STRING)},
-        returnType = {TypeEnum.INT},
+        returnType = {TypeEnum.DOUBLE},
         isPublic = true
 )
 @Component(
